requirements.txt
keyboard
numpy
opencv-python
Pillow
pypiwin32
